Ofertas_Laborales/requirements.txt
# ========================================
# DEPENDENCIAS DEL PROYECTO
# Sistema de Ofertas Laborales - Tacna
# ========================================

# Framework Web
Flask==2.3.3
Werkzeug==2.3.7

# Web Scraping
requests==2.31.0
beautifulsoup4==4.12.2
lxml==4.9.3

# Base de Datos NoSQL (MongoDB - NUEVO)
pymongo==4.6.1
dnspython==2.4.2  # Para conexiones MongoDB Atlas

# Base de Datos SQL (SQL Server - DEPRECADO, mantener para migración)
pyodbc==4.0.39

# Utilidades
python-dotenv==1.0.0

# Opcional: Para mejorar el scraping
selenium==4.16.0  # Para sitios con JavaScript dinámico
cloudscraper==1.2.71  # Para bypasear Cloudflare si es necesario
